deploy 1/8/22

+++ api/index.js
@@ -20,13 +20,18 @@
 const server = require('./src/app.js');
 const { conn } = require('./src/db.js');
 const apiInfo = require('./utils/apiCountries');
+require('dotenv').config()
 
-
+const port = process.env.PORT || 3001;
 // Syncing all the models at once.
-conn.sync({ force: true }).then(() => {
-  apiInfo().then(() => {
-    server.listen(3001, () => {
-      console.log('%s listening at 3001'); // eslint-disable-line no-console
-    });
-  })
-});
+if (process.env.NODE_ENV === 'development') {
+  conn.sync({ force: true }).then(() => {
+    apiInfo().then(() => {
+      server.listen(3001, () => {
+        console.log('%s listening at 3001'); // eslint-disable-line no-console
+      });
+    })
+  });
+} else {
+  server.listen(port, () => { console.log(`listening at ${port}`) })
+}
